add sqlite3 dependence

+++ src/distribution/distribution_test.cpp
@@ -18,7 +18,7 @@ void print_distribution(T distribution, const std::string& name) {
 
   std::cout << name << std::endl;
 
-  for (int i = 0; i < SLOT_NUM; ++i) {
+  for (size_t i = 0; i < SLOT_NUM; ++i) {
     std::cout << i << ": ";
     std::cout << std::string(p[i] * nstars / nrolls, '*') << std::endl;
   }
